+++ samples/Sample_01_SimpleText.php
@@ -4,7 +4,7 @@ include_once 'Sample_Header.php';
 // New Word Document
 echo date('H:i:s') , " Create new PhpWord object" , EOL;
 $phpWord = new \PhpOffice\PhpWord\PhpWord();
-$phpWord->addFontStyle('rStyle', array('bold' => true, 'italic' => true, 'size' => 16));
+$phpWord->addFontStyle('rStyle', array('bold' => true, 'italic' => true, 'size' => 16, 'allCaps' => true, 'doubleStrikethrough' => true));
 $phpWord->addParagraphStyle('pStyle', array('align' => 'center', 'spaceAfter' => 100));
 $phpWord->addTitleStyle(1, array('bold' => true), array('spaceAfter' => 240));
 
@@ -34,6 +34,7 @@ $fontStyle['strikethrough'] = true;
 $fontStyle['superScript'] = true;
 $fontStyle['color'] = 'FF0000';
 $fontStyle['fgColor'] = 'yellow';
+$fontStyle['smallCaps'] = true;
 $section->addText('I am inline styled.', $fontStyle);
 $section->addTextBreak();