+++ src/PhpWord/Writer/Word2007/Style/MarginBorder.php
@@ -9,8 +9,6 @@
 
 namespace PhpOffice\PhpWord\Writer\Word2007\Style;
 
-use PhpOffice\PhpWord\Settings;
-
 /**
  * Margin border style writer
  *
@@ -46,7 +44,6 @@ class MarginBorder extends AbstractStyle
     {
         $sides = array('top', 'left', 'right', 'bottom', 'insideH', 'insideV');
         $sizeCount = count($this->sizes) - 1;
-        $unit = Settings::getMeasurementUnit();
 
         for ($i = 0; $i < $sizeCount; $i++) {
             if (!is_null($this->sizes[$i])) {